make config available in correct place
