+++ core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/jndi/DirectorySubjectNameSupport.scala
@@ -1,5 +1,6 @@
 package org.broadinstitute.dsde.rawls.dataaccess.jndi
 
+import com.typesafe.config.ConfigFactory
 import org.broadinstitute.dsde.rawls.RawlsException
 import org.broadinstitute.dsde.rawls.model.{RawlsGroupName, RawlsUserSubjectId}
 
@@ -7,7 +8,15 @@ import org.broadinstitute.dsde.rawls.model.{RawlsGroupName, RawlsUserSubjectId}
   * Created by dvoet on 6/6/17.
   */
 trait DirectorySubjectNameSupport {
-  protected val directoryConfig: DirectoryConfig
+  private val conf = ConfigFactory.parseResources("version.conf").withFallback(ConfigFactory.load())
+
+  val directoryConfig = DirectoryConfig(
+    conf.getString("directory.url"),
+    conf.getString("directory.user"),
+    conf.getString("directory.password"),
+    conf.getString("directory.baseDn")
+  )
+
   val peopleOu = s"ou=people,${directoryConfig.baseDn}"
   val groupsOu = s"ou=groups,${directoryConfig.baseDn}"
