+++ core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/slick/DataAccessComponent.scala
@@ -5,5 +5,5 @@ import slick.driver.JdbcDriver
 
 import scala.concurrent.ExecutionContext
 
-class DataAccessComponent(val driver: JdbcDriver, val batchSize: Int, val directoryConfig: DirectoryConfig)(implicit val executionContext: ExecutionContext)
+class DataAccessComponent(val driver: JdbcDriver, val batchSize: Int, override val directoryConfig: DirectoryConfig)(implicit val executionContext: ExecutionContext)
 extends DriverComponent with DataAccess
